Restructured Angular Folders and Routes
