+++ src/main/front/src/app/services/animals.service.spec.ts
