+++ src/main/front/src/app/services/animals.service.ts
