+++ src/main/front/src/app/services/family.service.spec.ts
@@ -1,6 +1,6 @@
 import { TestBed } from '@angular/core/testing';
 
-import { FamilyService } from './family.service';
+import { FamilyService } from './services/family.service';
 
 describe('FamilyService', () => {
   let service: FamilyService;
